comment end

+++ spring_miniProject/spring_miniProject/src/main/java/com/hanghae/spring_miniProject/controller/PostController.java
@@ -3,6 +3,7 @@ package com.hanghae.spring_miniProject.controller;
 import com.hanghae.spring_miniProject.dto.FindAllPostRequestDto;
 import com.hanghae.spring_miniProject.dto.PostRequestDto;
 import com.hanghae.spring_miniProject.dto.PostResponseDto;
+import com.hanghae.spring_miniProject.dto.createPostResponseDto;
 import com.hanghae.spring_miniProject.model.Post;
 import com.hanghae.spring_miniProject.repository.PostRepository;
 import com.hanghae.spring_miniProject.security.UserDetailsImpl;
@@ -25,15 +26,15 @@ public class PostController {
 
     //게시글 등록
     @PostMapping("/api/post")
-    public PostResponseDto createPost(@RequestBody PostRequestDto requestDto, @AuthenticationPrincipal UserDetailsImpl userDetails){
+    public createPostResponseDto createPost(@RequestBody PostRequestDto requestDto, @AuthenticationPrincipal UserDetailsImpl userDetails){
         return postService.createPost(userDetails, requestDto);
     }
 
     //게시글 수정
     @PutMapping("/api/post/{postId}")
-    public String updatePost(@PathVariable Long postId, @RequestBody PostRequestDto requestDto) {
+    public String updatePost(@PathVariable Long postId, @RequestBody PostRequestDto requestDto, @AuthenticationPrincipal UserDetailsImpl userDetails) {
         try{
-            postService.update(postId, requestDto);
+            postService.update(postId, requestDto, userDetails);
             return "Success";
         }catch (Exception e){
             return e.getMessage();
@@ -42,12 +43,18 @@ public class PostController {
 
     //게시글 삭제
     @DeleteMapping("/api/post/{postId}")
-    public boolean DeletePost(@PathVariable Long postId){
+    public String DeletePost(@PathVariable Long postId, @AuthenticationPrincipal UserDetailsImpl userDetails){
         Post post = postRepository.findById(postId).orElseThrow(
                 () -> new NullPointerException("게시글이 존재하지 않습니다.")
         );
-        postRepository.deleteById(postId);
-        return true;
+
+        if(post.getUser().getId().equals(userDetails.getUser().getId())){
+            postRepository.deleteById(postId);
+        }
+        else {
+            return "본인의 작성한 게시글만 삭제할 수 있습니다.";
+        }
+        return "게시글 삭제를 완료하셨습니다.";
     }
 
     //게시글 전체 조회
